Added logging functionality
Since our VM keeps going down, and it is currently running in a screen session, I thought it would be a good idea to enable logging.

This way we don't have to keep looking at the console to debug.

The log should autorotate and keep 5 old versions, each about 1024000 bytes (I think)

I also added a parameter to name a log file location